+++ addons/website_twitter/models/__init__.py
@@ -1,2 +1,2 @@
-import twitter
-import twitter_config
+import website_twitter
+import website_twitter_config
